very large file with no newline causes trouble
grep doesn't deal well with very large files containing no line terminator.
I ran grep in a directory where I thought it'd find matches and complete
in a fraction of a second.  I was surprised to see it appear to hang and
finally exit with only this error: `grep: memory exhausted'.  The failure
was due to the presence of a file I'd created like this:

  dd bs=1 seek=1T of=big < /dev/null

So, even though that file occupies no disk blocks, grep thinks
it has to search a terabyte of data, and blithely tries to read
the first line (the entire file) into memory.

Of course this problem isn't unique to grep.
Any program that tries to read a line into memory at a time
is likely to meet the same fate.  But grep is a little different
from other tools in that one may well use it to search an entire
hierarchy for a file containing a particular string.  The mere
presence of a very large file should not cause such a search to fail.

Below, I've included a patch that makes grep report the error,
free the allocated memory, and continue with any remaining files.

Test like this:

  $ (ulimit -v 10000; echo a|./grep a big -)
  grep: big: Cannot allocate memory
  (standard input):a
  [Exit 2]

2003-04-03  Jim Meyering  <jim@meyering.net>

	When grep runs out of memory, don't abort the entire command,
	but rather just the affected command line argument(s).
	* src/grep.c (stdin_argv): New global.
	(fillbuf): Use malloc, not xmalloc, and handle malloc failure.
	(main): Rearrange main loop so that there is only one grepfile call.

Bug-Debian: http://bugs.debian.org/185208
Cc: bug-grep@gnu.org
Cc: bug-gnu-utils@gnu.org