-i doesn't work together with -o
The -o option causes patterns to be always matched
case-sensitively, regardless of -i.

Now, if the pattern contains uppercase chars the text is not
matched, and if the text contains uppercase chars they are
coerced to lowercase in the output!  (in a UTF-8 locale)

However, this seems to tickle a glibc bug [1], so make use of
RE_ICASE instead.

This fixes
    echo Y | LC_ALL=en_US.UTF-8 grep -i '[y]'
The expected output is:
    Y

Without this patch, it works on non UTF-8 environment, but fails on UTF-8
environment.

The definition of RE_ICASE comes from the glibc (/usr/include/regex.h)

Maybe lib/posix/regex.h should be removed to enforce the usage of the
glibc's regex.h

[1] <https://bugzilla.redhat.com/show_bug.cgi?id=112869>

Bug-Red-Hat: https://bugzilla.redhat.com/show_bug.cgi?id=111489
Bug-Debian: http://bugs.debian.org/181378