grep.1: use \(aq instead of '
grep's man page uses - instead of \- and ' instead of \(aq in
several places where these characters are intended to be used in
commands or (in one case) an email address.  On a UTF-8
configured system, this means that groff will get overfancy and
replace these characters with "better" ones which won't work when
pasted into a shell.

Bug-Debian: http://bugs.debian.org/226881