fix character ranges with grep -i in multibyte locales
On all locales that mb_cur_max > 1, such as CJK or UTF-8 locales,
[a-z] doesn't match with Z as expected if --ignore-case

For example,
 % echo Z | LANG=C grep --ignore-case '[a-z]'
 Z

 % echo Z | LANG=en_US.UTF-8 grep --ignore-case '[a-z]'
 %
 # wrong, Z should match [a-z] when --ignore-case

Note that [a-z] will match with A, that is not because --ignore-case works,
but because collation order in UTF-8 is "a A b B .. z".
That is, [a-z] won't match with Z even if --ignore-case.