Go faster in multibyte locales
The full story behind this patch is that grep-2.5.1a does not
handle UTF-8 gracefully at all. The basic plan with handling
UTF-8 in 2.5.1a is:

 * whenever a buffer is parsed, go through the entire buffer
   deciding how many bytes make up each character
 * use this information when necessary

This patch changes that to:

 * when information about how many bytes make up a character is
   needed, work it out on demand

On the face of it, this is a small obvious improvement. In fact
it is much better than that, because the original scheme would
calculate character lengths several times for each buffer: in
fact, one full pass for every single potential match!

More precisely, this reduces the multibyte processing (mbrtowc
etc) considerably by only using it when necessary.  For the
special case of UTF-8, without using the built-in DFA, this can
be never as far as grep is concerned; of course the system
re_search() function has to be aware of multibyte handling.

Bug-Red-Hat: https://bugzilla.redhat.com/show_bug.cgi?id=69900
Benchmarks: http://thread.gmane.org/gmane.comp.gnu.grep.bugs/100
Bug-Debian: http://bugs.debian.org/181378