grep -w in UTF-8 locale sometimes reports non-matches
The root cause of this bug is the use of mbtowc in 64-egf-speedup.patch
and 67-w.patch. These patches try to use mbtowc to look at the
character before and after the match to check if the match is a
whole word. But when a binary file is being grepped, mbtowc gets
passed random junk rather than a valid UTF-8 character. As a
consequence, its internal state gets messed up, and you get
nonsense for the following matches. The fix is to use mbrtowc so
you can reset its state. A patch is attached.

65-dfa-optional.patch is a red herring. I guess that patch just
exposes the bug because it causes grep to use a different code
path. But you get the same bug with grep -F, which is not touched
by that patch.

minimal test case:

echo -e '\344RC' > 0.bin
cat > 1.txt << EOF
RC
is not
ARCH
EOF

Testing this with

    LC_CTYPE=POSIX grep -w RC 0.bin 1.txt

yields the correct result

    0.bin:�RC
    1.txt:RC

while

    LC_CTYPE=en_US.UTF-8 grep -w RC 0.bin 1.txt

produces the invalid

    0.bin:�RC
    1.txt:RC
    1.txt:ARCH

Reported-by: arno renevier <arenevier@fdn.fr>
Testcase-by: Andreas Beckmann <debian@abeckmann.de>
Reviewed-by: Nicolas François <nicolas.francois@centraliens.net>
Bug-Debian: http://bugs.debian.org/446270