debian/rules: Migrate build target to use dh
This is a bit of an experiment: migrating debian/rules to the new
style one target at a time, so the changes for each can be
tested, tweaked, and reverted if necessary separately.

This should make debian/rules a little easier to read (because
shorter) and more familiar to those used to the more uniform
modern style for Debian source packages.