Add a list of generated files for use in the clean target
We do not use this in debian/rules clean yet, but we can soon.
Using '! test -e Makefile.in || make clean' is not very robust:
if the Makefile is deleted but other files have not been yet,
then there is no way to restart the build with a clean slate
short of regenerating the Makefile again.  Luckily, we do not
actually need the Makefile to delete all generated files:
upstream's .cvsignore files provide a perfectly adequate list
that can be used for that task.  One problem: the .cvsignore files
are not included in release tarballs.

So we add them to the debian directory.  Eventually, we should
add an autogen.sh script to regenerate this list so we can stop
tracking it in version control.